requirements.txt
pandas
numpy
requests
jupyter
matplotlib
seaborn
yfinance
riskfolio-lib
pyfolio
dotenv
# run 'conda install pysqlite3' for this to avoid errors